修改gis包bug，增加arcserver服务调用权限控制bug，增加坐标对批量导入支持;修改地区等级高后果区计算bug;

+++ src/main/java/cn/jasgroup/hcas/report/service/HcaReportService.java
@@ -327,7 +327,7 @@ public class HcaReportService extends CommonDataJdbcService {
 		}
 		NumberFormat numberFormat = NumberFormat.getInstance();
 		numberFormat.setMaximumFractionDigits(2);
-		String result = numberFormat.format((Double) (a / b) * 100) + "%";
+		String result = numberFormat.format(a / b * 100) + "%";
 		return result;
 	}
 }
